+++ GimmeMillions.Domain/Features/IFeatureExtractor.cs
@@ -7,6 +7,6 @@ namespace GimmeMillions.Domain.Features
     public interface IFeatureExtractor<TData>
     {
         string Encoding { get; }
-        float[] Extract(IEnumerable<(TData Data, float Weight)> data);
+        double[] Extract(IEnumerable<(TData Data, float Weight)> data);
     }
 }
