+++ GimmeMillions.Domain/ML/Binary/MLStockKernelEstimationSdcaModel.cs
@@ -101,7 +101,7 @@ namespace GimmeMillions.Domain.ML.Binary
             var inputDataView = _mLContext.Data.LoadFromEnumerable(
                 new List<StockRiseDataFeature>()
                 {
-                    new StockRiseDataFeature(input.Data, new float[0], false, 0.0f,
+                    new StockRiseDataFeature(input.Data, new double[0], false, 0.0f,
                     (int)input.Date.DayOfWeek / 7.0f, input.Date.Month / 366.0f)
                 },
                 GetSchemaDefinition(input));
@@ -168,7 +168,7 @@ namespace GimmeMillions.Domain.ML.Binary
                 {
                     var normVector = x.Input;
                     return new StockRiseDataFeature(
-                    normVector.Data, new float[0], x.Output.PercentDayChange >= 0,
+                    normVector.Data, new double[0], x.Output.PercentDayChange >= 0,
                     (float)x.Output.PercentDayChange,
                     (int)x.Input.Date.DayOfWeek / 7.0f, x.Input.Date.DayOfYear / 366.0f);
                 }),
